+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/repository/AccIdentityAccountRepository.java
@@ -1,5 +1,8 @@
 package eu.bcvsolutions.idm.acc.repository;
 
+import java.util.List;
+import java.util.UUID;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Modifying;
@@ -14,40 +17,24 @@ import eu.bcvsolutions.idm.core.api.repository.AbstractEntityRepository;
 import eu.bcvsolutions.idm.core.model.entity.IdmIdentity;
 
 /**
- * Target system configuration
+ * Identity accounts
  * 
  * @author Radek Tomiška
  *
  */
-//@RepositoryRestResource(//
-//		path = "identity-accounts", //
-//		collectionResourceRel = "identityAccounts", //
-//		itemResourceRel = "identityAccount", //
-//		exported = false // we are using repository metadata, but we want expose rest endpoint manually
-//	)
 public interface AccIdentityAccountRepository extends AbstractEntityRepository<AccIdentityAccount, IdentityAccountFilter> {
 	
-	/*
-	 * (non-Javadoc)
-	 * @see eu.bcvsolutions.idm.core.api.repository.BaseEntityRepository#find(eu.bcvsolutions.idm.core.api.dto.BaseFilter, Pageable)
+	/**
+	 * @deprecated Use AccIdentityAccountService (uses criteria api)
 	 */
 	@Override
-	@Query(value = "select e from AccIdentityAccount e left join e.identityRole ir" +
-	        " where" +
-	        " (?#{[0].accountId} is null or e.account.id = ?#{[0].accountId})" +
-	        " and" +
-	        " (?#{[0].identityId} is null or e.identity.id = ?#{[0].identityId})" +
-	        " and" +
-	        " (?#{[0].roleId} is null or ir.role.id = ?#{[0].roleId})" + 
-	        " and" +
-	        " (?#{[0].identityRoleId} is null or ir.id = ?#{[0].identityRoleId})" + 
-	        " and" +
-	        " (?#{[0].roleSystemId} is null or e.roleSystem.id = ?#{[0].roleSystemId})" + 
-	        " and" +
-	        " (?#{[0].systemId} is null or e.account.system.id = ?#{[0].systemId})" + 
-	        " and" +
-	        " (?#{[0].ownership} is null or e.ownership = ?#{[0].ownership})")
-	Page<AccIdentityAccount> find(IdentityAccountFilter filter, Pageable pageable);
+	@Deprecated
+	@Query(value = "select e from #{#entityName} e")
+	default Page<AccIdentityAccount> find(IdentityAccountFilter filter, Pageable pageable) {
+		throw new UnsupportedOperationException("Use AccIdentityAccountService (uses criteria api)");
+	}
+	
+	List<AccIdentityAccount> findAllByAccount_Id(UUID accountId);
 	
 	/**
 	 * Removes mapping by given account